lfs_project/requirements.txt
Django==1.7
django-model-utils==2.4
django-bootstrap-form==3.2
django-hitcount==1.2.1
django-passwords==0.3.7
django-precise-bbcode==1.2.2
django-annoying==0.9.0
django-form-utils==1.0.3
django-formset-js==0.5.0
django-jquery-js==2.1.4
pytz==2016.1
pybbm==0.17.3
Unidecode==0.4.19
bbcode==1.0.22
coverage==4.0.3
pbr==1.8.1
Pillow==3.0.0
six==1.10.0
stevedore==1.10.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
